requirements.txt
asgiref==3.8.1
coverage==7.6.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Django==5.1.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
django-cors-headers==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
celery==5.4.0
redis==5.0.8

beautifulsoup4==4.12.3 
opengraph_py3==0.71 
soupsieve==2.6

pytest==8.3.3
pytest-asyncio==0.21.0
pytest-django==4.9.0